Start loops

+++ Java/Day1/java2.java
@@ -7,11 +7,11 @@ public class java2 {
     public static void main (String[]arg){
         Scanner Ok = new Scanner(System.in);
         System.out.print("Enter Your First Number ");
-        int a = Ok.nextInt();
+        long a = Ok.nextInt();
         System.out.print("Enter Your 2nd Number ");
-        int b = Ok.nextInt();
+        long b = Ok.nextInt();
 
-        int sum = a + b ;
+        long sum = a + b ;
         System.out.print("Your sum is there ");
         System.out.println(sum);
